# app/static/swagger.yaml
swagger: "2.0"
info:
  title: Mechanic Shop API
  description: API documentation for the Mechanic Shop application.
  version: "1.0.0"
host: 127.0.0.1:5000
basePath: /
schemes:
  - http
paths:
  /customers/:
    get:
      tags:
        - Customers
      summary: "Get all customers"
      description: "Returns a paginated list of customers."
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
          description: "Page number. Defaults to 1."
        - in: "query"
          name: "per_page"
          type: "integer"
          required: false
          description: "Number of customers per page. Defaults to 2."
      responses:
        200:
          description: "Customers retrieved successfully."
          schema:
            $ref: "#/definitions/CustomerListResponse"
          examples:
            application/json:
              generated_at: "2026-09-14T08:00:00"
              customers:
                - id: 1
                  first_name: "Tim"
                  last_name: "Sussmann"
                  email: "customer@example.com"
                  phone: "555-123-4567"
                  address: "123 Main Street"
              page: 1
              per_page: 2
              total: 1
              pages: 1

    post:
      tags:
        - Customers
      summary: "Create a new customer"
      description: "Creates a new customer account."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully."
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 2
              first_name: "John"
              last_name: "Doe"
              email: "john.doe@example.com"
              phone: "555-987-6543"
              address: "456 Oak Street"
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: "Get customer's service tickets"
      description: "Returns a list of service tickets belonging to the authenticated customer."
      security:
        - Bearer: []
      responses:
        200:
          description: "Customer's service tickets retrieved successfully."
          schema:
            $ref: "#/definitions/MyTicketsResponse"
          examples:
            application/json:
              - id: 1
                customer_id: 2
              - id: 2
                customer_id: 2
        401:
          description: "Authentication required."
        404:
          description: "Customer not found."  
  /customers/{id}:
    put:
      tags:
        - Customers
      summary: "Update a customer"
      description: "Updates one or more fields for an existing customer."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          required: true
          description: "Customer ID."
        - in: "body"
          name: "body"
          description: "Customer fields to update."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully."
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              first_name: "Tim"
              last_name: "Sussmann"
              email: "tim@example.com"
              phone: "555-123-4567"
              address: "123 Main Street"
        404:
          description: "Customer not found." 

    delete:
      tags:
        - Customers
      summary: "Delete a customer"
      description: "Deletes an existing customer."
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          required: true
          description: "Customer ID."
      responses:
        200:
          description: "Customer deleted successfully."
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer deleted successfully"
        404:
          description: "Customer not found."
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Customer login"
      description: "Authenticates a customer using an email address and password."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer login credentials."
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful."
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        401:
          description: "Invalid credentials."
          schema:
            $ref: "#/definitions/LoginErrorResponse"
          examples:
            application/json:
              message: "Invalid credentials"
...
definitions:
  CreateCustomerPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
      address:
        type: "string"
      password:
        type: "string"
        format: "password"
    required:
      - first_name
      - last_name
      - email
      - phone
      - address
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
      address:
        type: "string"
  CustomerListResponse:
    type: "object"
    properties:
      generated_at:
        type: "string"
      customers:
        type: "array"
        items:
          $ref: "#/definitions/CreateCustomerResponse"
      page:
        type: "integer"
      per_page:
        type: "integer"
      total:
        type: "integer"
      pages:
        type: "integer"
  UpdateCustomerPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
      address:
        type: "string"
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
        format: "email"
      phone:
        type: "string"
      address:
        type: "string"

    DeleteCustomerResponse:
      type: "object"
      properties:
        message:
          type: "string"

    LoginCredentials:
      type: "object"
      properties:
        email:
          type: "string"
          format: "email"
        password:
          type: "string"
          format: "password"
        required:
          - email
          - password

    LoginResponse:
      type: "object"
      properties:
        token:
          type: "string"

    LoginErrorResponse:
      type: "object"
      properties:
        message:
          type: "string"
  MyTicketsResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        customer_id:
          type: "integer"

    CreateMechanicPayload:
      type: "object"
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
          format: "email"
        phone:
          type: "string"
      required:
        - first_name
        - last_name
        - email
        - phone

    CreateMechanicResponse:
      type: "object"
      properties:
      id:
        type: "integer"
      first_name:
        type: "string"
      last_name:
        type: "string"
      email:
        type: "string"
        format: "email"
        phone:
          type: "string"

    MechanicListResponse:
      type: "array"
      items:
        type: "object"
        properties:
          id:
            type: "integer"
          first_name:
            type: "string"
          last_name:
            type: "string"
          email:
            type: "string"
            format: "email"
          phone:
            type: "string"
          ticket_count:
            type: "integer"

    UpdateMechanicPayload:
      type: "object"
      properties:
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
          format: "email"
        phone:
          type: "string"

    UpdateMechanicResponse:
      type: "object"
      properties:
        id:
          type: "integer"
        first_name:
          type: "string"
        last_name:
          type: "string"
        email:
          type: "string"
          format: "email"
        phone:
          type: "string"

    DeleteMechanicResponse:
      type: "object"
      properties:
        message:
          type: "string"

    CreateInventoryPayload:
      type: "object"
      properties:
        name:
          type: "string"
        price:
          type: "number"
      required:
        - name
        - price

    CreateInventoryResponse:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        price:
          type: "number"

    InventoryListResponse:
      type: "array"
      items:
        $ref: "#/definitions/CreateInventoryResponse"

    UpdateInventoryPayload:
      type: "object"
      properties:
        name:
          type: "string"
        price:
          type: "number"

    UpdateInventoryResponse:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        price:
          type: "number"

    DeleteInventoryResponse:
      type: "object"
      properties:
        message:
          type: "string"

    CreateServiceTicketPayload:
      type: "object"
      properties:
        customer_id:
          type: "integer"
      required:
        - customer_id

    CreateServiceTicketResponse:
      type: "object"
      properties:
        id:
          type: "integer"
        customer_id:
          type: "integer"
        inventory:
          type: "array"
          items:
            $ref: "#/definitions/CreateInventoryResponse"

    ServiceTicketListResponse:
      type: "array"
      items:
        $ref: "#/definitions/CreateServiceTicketResponse"

    EditServiceTicketPayload:
      type: "object"
      properties:
        add_ids:
          type: "array"
          items:
            type: "integer"
        remove_ids:
          type: "array"
          items:
            type: "integer"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header